api: Move api under v1
Move the api under as v1.
The wake action is now under /api/v1/wake.
